pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rxgym"
version = "1.0.0"
description = "OpenEnv clinical pharmacology environment for dose finding, allometric scaling, and DDI-aware scheduling."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "openenv-core>=0.2.3",
  "fastapi>=0.135.3",
  "uvicorn>=0.43.0",
  "pydantic>=2.12.5",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
py-modules = ["__init__", "client", "models", "inference", "tasks"]

[tool.setuptools.packages.find]
where = ["."]
include = ["server", "rl_agent", "rl_agent.*"]
